+++ src/RcppExports.cpp
@@ -240,6 +240,18 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// get_capacity_selected_
+int get_capacity_selected_(SEXP sim_, int id);
+RcppExport SEXP _simmer_get_capacity_selected_(SEXP sim_SEXP, SEXP idSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type sim_(sim_SEXP);
+    Rcpp::traits::input_parameter< int >::type id(idSEXP);
+    rcpp_result_gen = Rcpp::wrap(get_capacity_selected_(sim_, id));
+    return rcpp_result_gen;
+END_RCPP
+}
 // get_queue_size_
 int get_queue_size_(SEXP sim_, const std::string& name);
 RcppExport SEXP _simmer_get_queue_size_(SEXP sim_SEXP, SEXP nameSEXP) {
@@ -249,6 +261,18 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< SEXP >::type sim_(sim_SEXP);
     Rcpp::traits::input_parameter< const std::string& >::type name(nameSEXP);
     rcpp_result_gen = Rcpp::wrap(get_queue_size_(sim_, name));
+    return rcpp_result_gen;
+END_RCPP
+}
+// get_queue_size_selected_
+int get_queue_size_selected_(SEXP sim_, int id);
+RcppExport SEXP _simmer_get_queue_size_selected_(SEXP sim_SEXP, SEXP idSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type sim_(sim_SEXP);
+    Rcpp::traits::input_parameter< int >::type id(idSEXP);
+    rcpp_result_gen = Rcpp::wrap(get_queue_size_selected_(sim_, id));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -264,6 +288,18 @@ BEGIN_RCPP
     return rcpp_result_gen;
 END_RCPP
 }
+// get_server_count_selected_
+int get_server_count_selected_(SEXP sim_, int id);
+RcppExport SEXP _simmer_get_server_count_selected_(SEXP sim_SEXP, SEXP idSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type sim_(sim_SEXP);
+    Rcpp::traits::input_parameter< int >::type id(idSEXP);
+    rcpp_result_gen = Rcpp::wrap(get_server_count_selected_(sim_, id));
+    return rcpp_result_gen;
+END_RCPP
+}
 // get_queue_count_
 int get_queue_count_(SEXP sim_, const std::string& name);
 RcppExport SEXP _simmer_get_queue_count_(SEXP sim_SEXP, SEXP nameSEXP) {
@@ -273,6 +309,18 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< SEXP >::type sim_(sim_SEXP);
     Rcpp::traits::input_parameter< const std::string& >::type name(nameSEXP);
     rcpp_result_gen = Rcpp::wrap(get_queue_count_(sim_, name));
+    return rcpp_result_gen;
+END_RCPP
+}
+// get_queue_count_selected_
+int get_queue_count_selected_(SEXP sim_, int id);
+RcppExport SEXP _simmer_get_queue_count_selected_(SEXP sim_SEXP, SEXP idSEXP) {
+BEGIN_RCPP
+    Rcpp::RObject rcpp_result_gen;
+    Rcpp::RNGScope rcpp_rngScope_gen;
+    Rcpp::traits::input_parameter< SEXP >::type sim_(sim_SEXP);
+    Rcpp::traits::input_parameter< int >::type id(idSEXP);
+    rcpp_result_gen = Rcpp::wrap(get_queue_count_selected_(sim_, id));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -1157,9 +1205,13 @@ static const R_CallMethodDef CallEntries[] = {
     {"_simmer_get_attribute_", (DL_FUNC) &_simmer_get_attribute_, 3},
     {"_simmer_get_prioritization_", (DL_FUNC) &_simmer_get_prioritization_, 1},
     {"_simmer_get_capacity_", (DL_FUNC) &_simmer_get_capacity_, 2},
+    {"_simmer_get_capacity_selected_", (DL_FUNC) &_simmer_get_capacity_selected_, 2},
     {"_simmer_get_queue_size_", (DL_FUNC) &_simmer_get_queue_size_, 2},
+    {"_simmer_get_queue_size_selected_", (DL_FUNC) &_simmer_get_queue_size_selected_, 2},
     {"_simmer_get_server_count_", (DL_FUNC) &_simmer_get_server_count_, 2},
+    {"_simmer_get_server_count_selected_", (DL_FUNC) &_simmer_get_server_count_selected_, 2},
     {"_simmer_get_queue_count_", (DL_FUNC) &_simmer_get_queue_count_, 2},
+    {"_simmer_get_queue_count_selected_", (DL_FUNC) &_simmer_get_queue_count_selected_, 2},
     {"_simmer_Seize__new", (DL_FUNC) &_simmer_Seize__new, 5},
     {"_simmer_Seize__new_func", (DL_FUNC) &_simmer_Seize__new_func, 5},
     {"_simmer_SeizeSelected__new", (DL_FUNC) &_simmer_SeizeSelected__new, 5},