+++ tests/testthat/test-trajectory-set-capacity-queue.R
@@ -49,14 +49,26 @@ test_that("capacity and queue size change as expected (1)", {
 
 test_that("capacity and queue size change as expected (2)", {
   t <- trajectory() %>%
+    seize("dummy0", 2) %>%
+    seize("dummy1", 3) %>%
     select("dummy0", id = 0) %>%
     select(function() "dummy1", id = 1) %>%
     set_capacity_selected(0, id = 0) %>%
     set_capacity_selected(function() Inf, id = 1) %>%
     set_queue_size_selected(Inf, id = 1) %>%
-    set_queue_size_selected(function() 5, id = 0)
-
-  expect_output(print(t))
+    set_queue_size_selected(function() 5, id = 0) %>%
+    log_(function() paste(
+      get_capacity_selected(env, id=0),
+      get_server_count_selected(env, id=0),
+      get_queue_size_selected(env, id=0),
+      get_queue_count_selected(env, id=0),
+      get_capacity_selected(env, id=1),
+      get_server_count_selected(env, id=1),
+      get_queue_size_selected(env, id=1),
+      get_queue_count_selected(env, id=1)
+    )) %>%
+    release("dummy0", 2) %>%
+    release("dummy1", 3)
 
   env <- simmer(verbose = TRUE) %>%
     add_resource("dummy0", 3, 0) %>%
@@ -67,7 +79,7 @@ test_that("capacity and queue size change as expected (2)", {
   expect_equal(env %>% get_capacity("dummy1"), 3)
   expect_equal(env %>% get_queue_size("dummy0"), 0)
   expect_equal(env %>% get_queue_size("dummy1"), 0)
-  env %>% stepn(4)
+  env %>% stepn(6)
   expect_equal(env %>% get_capacity("dummy0"), 0)
   env %>% stepn()
   expect_equal(env %>% get_capacity("dummy1"), Inf)
@@ -75,6 +87,7 @@ test_that("capacity and queue size change as expected (2)", {
   expect_equal(env %>% get_queue_size("dummy1"), Inf)
   env %>% stepn()
   expect_equal(env %>% get_queue_size("dummy0"), 5)
+  expect_output(run(env), "0 2 5 0 Inf 3 Inf 0")
 })
 
 preempt <- trajectory() %>% set_capacity("res", 0)