+++ NEWS.md
@@ -4,6 +4,7 @@
 
 * New `timeout_from_attribute()` activity makes it easier to set a timeout based on an attribute (#129).
 * The activities `set_attribute()`, `set_prioritization()`, `set_capacity()` and `set_queue_size()` get a new argument `mod` which, if set to `"+"` or `"*"`, modifies the corresponding value instead of substituting it. This makes it easier to increment, decrement or scale one of these values (#130).
+* New `*_selected()` versions for the already available resource getters: `get_capacity()`, `get_queue seize()`, `get_server_count()` and `get_queue_count()` (#134).
 
 ## Minor changes and fixes: